add tests, several other fixes
