+++ examples/ecoli-reads.png
