+++ examples/hist.png
