--- tests/test_sourmash_plugin.py
@@ -1,18 +1,0 @@
-"""
-Tests for sourmash_plugin_xyz.
-"""
-import os
-import pytest
-
-import sourmash
-import sourmash_tst_utils as utils
-from sourmash_tst_utils import SourmashCommandFailed
-
-
-def test_run_sourmash(runtmp):
-    with pytest.raises(SourmashCommandFailed):
-        runtmp.sourmash('', fail_ok=True)
-
-    print(runtmp.last_result.out)
-    print(runtmp.last_result.err)
-    assert runtmp.last_result.status != 0                    # no args provided, ok ;)
